requirements.txt
django==5.0.1
stripe==7.11.0
djangorestframework
python-dotenv
